Reorganize endpoint import to prevent circular import
