+++ backend/app/app/api/api_v1/api.py
@@ -6,6 +6,7 @@ api_router = APIRouter()
 api_router.include_router(endpoints.login.router, tags=["login"])
 api_router.include_router(endpoints.users.router, prefix="/users", tags=["users"])
 api_router.include_router(endpoints.utils.router, prefix="/utils", tags=["UI"])
+api_router.include_router(endpoints.scenario.router, prefix="/utils", tags=["UI"])
 api_router.include_router(endpoints.ue_movement.router, prefix="/ue_movement", tags=["Movement"])
 api_router.include_router(endpoints.paths.router, prefix="/paths", tags=["Paths"])
 api_router.include_router(endpoints.gNB.router, prefix="/gNBs", tags=["gNBs"])
