+++ backend/app/app/api/api_v1/endpoints/__init__.py
@@ -20,3 +20,4 @@ from .racsProvisioning import router
 from .analyticsExposure import router
 from .broker import router
 from .tests import router
+from .scenario import router
